+++ src/lib/FlightTasks/tasks/AutoMapper2/FlightTaskAutoMapper2.hpp
@@ -41,6 +41,7 @@
 #pragma once
 
 #include "FlightTaskAuto.hpp"
+#include "UtilityFunctions.hpp"
 
 class FlightTaskAutoMapper2 : public FlightTaskAuto
 {
@@ -77,5 +78,4 @@ private:
 
 	void _reset(); /**< Resets member variables to current vehicle state */
 	WaypointType _type_previous{WaypointType::idle}; /**< Previous type of current target triplet. */
-	bool _highEnoughForLandingGear(); /**< Checks if gears can be lowered. */
 };