+++ src/lib/FlightTasks/tasks/Utility/UtilityFunctions.cpp
@@ -1,0 +1,12 @@
+/**
+ * @file UtilityFunctions.cpp
+ */
+
+#include "UtilityFunctions.hpp"
+
+bool highEnoughForLandingGear(const float alt_threshold, const float reference_alt)
+{
+
+	return  reference_alt > alt_threshold;
+
+}