+++ src/lib/FlightTasks/tasks/Utility/UtilityFunctions.hpp
@@ -1,0 +1,10 @@
+/**
+ * @file UtilityFunctions.hpp
+ *
+ * This file is used to some utility functions.
+ */
+
+#pragma once
+
+/** Return true if the current height allows landing gear operation */
+bool highEnoughForLandingGear(const float alt_threshold, const float reference_alt);